requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
anthropic==0.40.0
supabase==2.10.0
yfinance==0.2.50
pandas==2.2.3
numpy==1.26.4
httpx==0.27.2
aiohttp==3.11.10
apscheduler==3.10.4
feedparser==6.0.11
python-multipart==0.0.20
pydantic-settings==2.6.1
pydantic==2.10.3
